Allow skipping of automatic region replacement
If you define in serverless.yaml:

    custom:
       pseudoParameters:
          skipRegionReplace: true

You can skip the automatic replacement of anything that looks like a
region with AWS::Region.

This is necessary for example when working with Cloudfront. Cloudfront
only works with ACM certificates that are in us-east-1. So if we replace
with anything else (e.g us-west-2) you will break the cloudfront distribution.